Extract callbacks to a new module
